// src/shared/api/models/getApiTax200Item.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiTax200ItemApplyTaxTo } from './getApiTax200ItemApplyTaxTo';
import type { GetApiTax200ItemTaxType } from './getApiTax200ItemTaxType';
import type { GetApiTax200ItemCreatedAt } from './getApiTax200ItemCreatedAt';
import type { GetApiTax200ItemUpdatedAt } from './getApiTax200ItemUpdatedAt';

export type GetApiTax200Item = {
    id: string;
    /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
    seqId: number;
    /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
    outletId: number;
    taxName: string;
    /**
     * @minimum -140737488355328
     * @maximum 140737488355327
     */
    taxRate: number;
    applyTaxTo: GetApiTax200ItemApplyTaxTo;
    taxType: GetApiTax200ItemTaxType;
    taxOnCustomAmounts: boolean;
    taxExemption: boolean;
    includeTaxInItemPrice: boolean;
    isActive: boolean;
    createdAt: GetApiTax200ItemCreatedAt;
    updatedAt: GetApiTax200ItemUpdatedAt;
};